Concluding this project
I have created the my books section.
This project was really helpful for learning backend and networking concepts .....
front end is a lot messier then backend because of my lack of front end expertise.
I still tried to keep things fucntional and focus on more important concept that is practicing the core concepts ...